Rewrite code to implement new logic
Previous version was determining what has changed in a source directory
by comparing its two snapshots made at different time points.  Creating
a snapshot relied heavily on stat() function which might lead to
"beating down" the file system during normal operations due to a
combination of the number of files in the monitored directory and/or
sampling frequency.

However, in the current design of the system:

1. Writing files to the buffer is almost instantaneous (The Archiver
   does not actually copy the files to the buffer, but creates hard
   links instead.
2. Files in the buffer are immutable (The link is created after any
   writing to the actual file is completed).
3. After successful transfer, files are moved from the buffer to the
   holding area (a different directory).

As a result, the program could be modified to use a much simpler logic
as the buffer either contain new files or files which were not
transferred successfully in the past and need to be transferred anyway.
That eliminated the need to create and maintained directory snapshots.