Rewrite Scanner and Porter as commands
Two core classes: Scanner and Porter were subclasses of the Thread
class.  However, they were not run as daemons. As a result they needed
to be created and initialized in each iteration of the main loop.  I
rewrote them as commands. That way they can be initialized just once and
then run repeatedly, possibly in threads.