Make sure file transfer is "atomic"
In its first iteration, ctrl_oods will be used to emulate ingest
service.  However, it was relies on the fact that that placing files in
the OODS buffer is an atomic operation.  To emulate such a behavior I
modified the code to initially transfer file to a temporary staging area
from which they are moved (once transfer is completed) the buffer
managed by ctrl_oods.  Also added some cleanup commands to keep buffer
and staging area free from "leftover" directories.