Run cleanup in the background
Files which were successfully transferred are moved from the buffer to
the holding area.  In previous implementation they were moved after all
files scheduled for transfer were processed.  This could lead to a
significant inconsistency with regard to files in the buffer and files
which were already transferred in case of the program crashing during
processing of large amount of files.  To mitigate this risk I modified
the program to run the cleanup in a background thread and move the
successfully transferred files to the holding area practically
immediately.