Change buffer cleanup policy
Directories in the buffer of the handoff site were removed as soon as
they became empty.  This policy could lead to a race condition between
the archiver creating a directory to write a file and the buffer manager
removing it.  I modified the cleanup policy to remove empty
directories in the buffer only if no file was written to them for a
a given period of time.