Call Wiper() unconditionally after transfer
Porter threads populate a queue with the names of the successfully
transferred files.  The size of the queue was used to determine if the
command removing possible empty directories on the endpoint site should
be run.  However, the same queue is being drained by the background
thread moving files from the buffer to the holding area.  It renders
size of the queue rather poor indicator if any files were transferred.
Hence, I modified to code to always call the clean-up command on the
endpoint site after all files are processed.