Move files directly to buffer
Using 'ctrl_oods' as a substitute for a real endpoint buffer manager
required handoff buffer manager to place a successfully transferred file
in the storage area and create a hard link it in the buffer as the
'ctrl_oods' was moving files to the GenX Butler repository during the
ingestion process.  With actual endpoint manager, it is no longer
necessary. Now it is the endpoint manager responsibility to move the
file from the buffer to the storage area. On top of that making copies
is no longer necessary as by default the endpoint manager uses linking
when ingesting files to a GenX Butler repository instead of moving them.
Hence, I altered the file transfer process in the handoff manager
accordingly.