Collect metrics about files and transfer attempts
The handoff manager was not storing any data about files in the buffer
and transfer attempts it made.  As a result, it was difficult address
questions like: when a file was created, was it transferred yet, what
was the transfer rate, etc.  To address this limitation I modified the
manager to maintain a SQLite database with various pieces of information
that an operator or a monitoring system could relatively easily query to
find answers for the question mentioned earlier.