Do not move files out of buffer if transfer failed
After creating database entries corresponding to transfer batches, the
manager was passing "downstream" file messages even for files that were
not transferred to the endpoint site.  These files were being moved out,
incorrectly, of the buffer to the holding area. As a result, no further
transfer attempts could be made for these files.  I changed the code to
make sure that only messages for successfully transferred files will be
created.