Make creation of the Batch record more robust
Some fields in the message describing a transfer batch needs to be
converted to different type when the corresponding database record is
created.  These conversion were done unconditionally causing issues when
the fields were not set (equaled None).  To address these, I created a
function which converts the transfer massage to a database record
handling gracefully such corner cases.