Stop manager from crushing during network outages
A wrong datatype was used when the manager was populating transfer
message's fields related to its post-transfer activities. Specifically,
it was assigning datetime.timedelta() instead of a float representing
the time it took to create a directory on the remote site.  As a result,
the manager was prone to crashing during the network outages as on rare
occasions it was attempting to convert said value to
datetime.timedelta() again when persisting the message in its database.
Fixed the problem by assigning a correct type during message creation.